fix(logscrub,logattrs): ASCII-only fold in scrub regexes; exact 64-bit ints
logscrub: the built-in patterns used (?i), which Go folds via
unicode.SimpleFold — `(?i)password` also matches `paſsword` (U+017F LATIN
SMALL LETTER LONG S) and `(?i)token` matches `toKen` (U+212A KELVIN SIGN) —
while the literal prefilters that gate them fold ASCII only. Such a line was
rejected by the prefilter, so the pattern never ran and the secret shipped
UNREDACTED: a security control quietly narrower than the regex it
advertises. Spell the keywords with explicit ASCII case classes so the
regex's reach equals the prefilter's, which is what makes the prefilter
optimisation sound. The new test asserts that equality for every built-in
rather than just the cases we thought of. BenchmarkScrubNoMatch stays at
0 allocs/op.

logattrs: JSON numbers all decoded through ParseFloat, so an integer above
2^53 lost precision — a 64-bit id (snowflake, order/user id) lifted from a
structured log silently landed one or more off, and still looked exact
downstream because apply.Put stores whole floats with PutInt. Parse integral
tokens as int64 first; apply.Put and the grouping key already honor int64,
so the exported attribute is unchanged for values that fit either way.

+++ pkg/logattrs/logattrs_test.go
@@ -40,7 +40,10 @@ func TestExtractJSON(t *testing.T) {
 	for _, a := range r.Log {
 		got[a.Key] = a.Val
 	}
-	if got["level"] != "warn" || got["status"] != float64(503) || got["cached"] != true || got["ratio"] != 0.5 {
+	// status is an INTEGRAL JSON number, so it decodes as int64 (float64 cannot
+	// hold a 64-bit id exactly); ratio keeps float64. Either way apply.Put
+	// stores a whole number with PutInt, so the exported attribute is the same.
+	if got["level"] != "warn" || got["status"] != int64(503) || got["cached"] != true || got["ratio"] != 0.5 {
 		t.Errorf("log = %+v", got)
 	}
 	if _, ok := got["absent"]; ok {
